[auto] various improvements

+++ adaptive-lib-auto/src/commonMain/kotlin/fun/adaptive/auto/model/PeerId.kt
@@ -1,3 +1,3 @@
 package `fun`.adaptive.auto.model
 
-typealias PeerId = Int
+typealias PeerId = Long
